python-mqtt-kafka/requirements.txt
flask==3.0.0
pypika==0.48.9
mysql-connector-python==8.2.0
bcrypt==4.0.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
argparse==1.4.0
hrv-analysis==1.0.4
Flask-MQTT==1.1.0
kafka-python==2.0.2
